DCS-203: Bass journey changed to accept Approved Address
http://localhost:3000/hdc/bassReferral/bassAreaCheck/
http://localhost:3000/hdc/bassReferral/approvedPremisesAddress/
http://localhost:3000/hdc/bassReferral/approvedPremisesChoice/

Previously if the prisoner requested a Bass address, the RO didn't have the ability to change it.The RO can now override the Bass request to an Approved Premises (AP) address.

The additional journey has resulted in new logic to manage the tasklists and caselist statuses.

The option to use an AP address begins in the Bass Area Check screen where a new radio button group has been added asking if an AP is required. The two options are Yes and No.
If Yes is seleted, the RO is taken to the AP address screen where he can enter the address details. If No is selected, the user is returned to the tasklist screen

The new journey was made to be similar to the journey where a user requested his own Curfew Address but instead the RO decided an AP was appropriate. In that journey there is an intermediary page (after the RO's intervention) where the CA  c request AP address. A similar intermediary page (hdc/bassReferral/approvedPremisesChoice) has been added to this new BASS -> AP address journey. This intermediary page is presented when the CA selects the 'view/edit' button against the Bass Address task in the Tasklist.

Added new validation for approvedPremisesAddress
Added error handling in bassAreaCheck.pug'
Updated the current unit tests to account for the new code. Specifically, added the approvedPremisesRequiredYesNo to the object for the Bass address screen.

Caselist and tasklist labels and statuses displayed have been updated when RO chooses  Approved Premises address even though prisoner may have requested Bass accommodation.

The status of Opted Out will show if user decides to opt out.