DCS-233 Calling delius when staff doesn't exist.
Merely logging result for now.

Also throwing error if can't get a token using client credentials when calling delius.

This would have previously thrown an auth error which would log the user out of the app but as this is client credential this situation would only happen with a programming error (so should result in a 500 status)