DCS-276: Create caService
New service to determine whether the application can progress from CA to RO during the eligibility stage.
The service uses the offenders bookingId to find the RO. The RO can either be an Prison Offender Manager (POM) or Community Offender Manager (COM).

We determine the LDU via the RO. The RO to LDU relationship is defined in Delius.
We check the LDU returned from Delius against the activeLdu's DB table. If the LDU is active, then it satisfies another condition for progression.

The caService itself will return either error codes if the CA should be prevented from progressing the case to the RO or a null if there are no blockers.

The happy path for progression from CA to RO is
	1, the RO is a COM
	2, the LDU can be determined
	3, the LDU is active (i.e in the geo areas that are being rolled out)

DCS-276: pinpointing build errors by commenting parts of caServiceTest.js