DCS-262 Recording warnings
Going to provide a general error/warnings queue to expose to the internal team.
This ticket covers storeing warnings whenever a CA attempts to assign a case to an RO who has an unlinked account.
These ROs will not be notified that a case has been assigned to them - and some manual work needs to be done to fix the record.

There are other places where we merely log errors on things which end up biting us which could be surfaced in the queue and reduce the maintenance overhead of the service.