DCS-264: CA tasklist displays any error messages and also disables continue button
http://localhost:3000/hdc/taskList/bookingId

The CA is prevented from progressing from from the tasklist if there are any errors during the eligibility stage.
This means the CA will not be able to complete the curfew address or handoff to the RO

The errors can result from the info in delius not being consistent or the  LDU not beingactive.

The tasklist will display the error messages in a red border, similar to the GDS style but they won't be at the top of the page or be hyperlinked.

The Start now button is disabled if any errors are identified.