DCS-264-feature flag created and tests updated
The feature flag will enable the user to either follow the happy path where the CA can complete the eligibility section and then enter the Curfew Address or alternatively, be prevented from progressing past eligibilty checks if the ROs LDU is not active.
The default for the flag is to let the CA progress to the Curfew Address section.