DCS-265: new column in activeLdu table
Inserted a new column in active_local_delivery_units table

The new column is probabtion_area_code.
This is needed to link an LDU to a probation area
Updated the addLdu and isLduPresent function in the server/data/activeLduClient.js file as the save and search they perform need to include the probabtion area.