DCS-267 Adding migration to activate all LDUs in the north east.
The north east is already live with HDC and this is to ensure that we continue to allow CAs to handover cases to ROs in the north east once we enable the CA blocking feature.

Also removing unused type definition which described the return type of a now unused delius api call.