DCS 265 new UI for Probation and Ldu pages
Two new pages added:
1. Probation areas: http://localhost:3000/admin/locations/probation-areas
This displays a list of hyperlinks for all the probation areas currently in Delius. Selecting a hyperlink takes the user to the next page that displays all the LDUs associated with that probation area. The url segment passes forward the code for the probation area . This list appears in alpha order.
2. Local Delivery Units page:
http://localhost:3000/admin/locations/probation-areas/Bir/local-delivery-units
Displays all the LDUs associated with a specific probation area. This page displays a checkbox for each LDU.
If there are no LDUs' to display, an information message is displayed.
The checkboxes will be preselected if that LDU is in the active_ldu table of the Licences DB.
Selecting or de-selecting a checkbox followed by save will update the DB.
The Probation area page is accessed via http://localhost:3000/admin/