DCS 318: new licence additional condition added to the drugs section
A new condition asking about drugs dependency added to the Drugs, Health and Behaviour section.
The new content is the same as that at the bottom of the page under the Post-sentance section.
The printed licences have also been updated so that the new content appears in either the main section or the PSS section, depending on the licence type