DCS-331 backend work to support admin staff sending emails to ROs
This will allow a frontend to be developed to allow sending emails to new linked RO users.
The original work to transition the state and reset any decisions will have already occurred when the CA handed over the case.

The original handover will have triggered a SEND audit event using the PCA username.
This will trigger a NOTIFY event with the admin staffs username.
It will also attempt to assign the HDC Update role to the RO's delius user.

This new service method will return the RO to provide some contextual information on success.
Otherwise it will return the error for display to the user.
Unlike the existing service method it will error if the user is not linked - the assumption will
be that this will only be called once the user is linked