DCS-358 Raising app insights event when case is handed over.
This event will contain information about the source and target of the case.
This should allow us to easily see frequency at which certain prison/LDUs handover cases.

Also adding this information to the audit trail in case we want to surface this through the
app at later date.

Whilst doing this I noticed that we weren't capturing any audit info for DMs - we're now doing that.