DCS-360- error/warning messages no longer displayed in DM tasklist page
http://localhost:3000/hdc/taskList/{bookingId}

Changes made in server/routes/curfew.js . If a DM refuses a curfew address and hands the case back to the CA, the CA can withdraw the address and submit a new one.  When the CA hands back to the DM, a warning message displays unnecessarily  in the DM tasklist. To prevent the warnings from displaying, have removed the 'approval' object in the licence data in the DB because it holds details of the DM's previous decision. Achieved by executing licenceService.removeDecision(bookingId, res.locals.licence) in curfew.js
The licenceService.getLicence(bookingId)  is then executed so that subsequent code has  the updated version of the licence data.