DCS-384 Fixing issue with bad request when getting sentences from elite
If we don't specify offender numbers when requesting sentences, elite throws a 400 bad request.
We had a guard to protect against this but it assumed that offender numbers would always be
present in delius for each prisoner and this is not always the case.

Offenders without offender numbers would be filtered out of the request but
If all prisoners did not have offender numbers then this would result in a 400

This fixes the guard to prevent this edgecase.