DCS-271: CA can resubmit to DM
CA can resubmit the case back to DM from the CA tasklist. The new task on the tasklist only displays if the DM has already made a decision (approval or rejection). It does not appear in the tasklist prior to DM that point.
Updated failing unit tests for CA tasklist to take new task into account.