DCS-411: changes to HDC refusal message (#1163)
* DCS-411: changes to HDC refusal message
This is temporary fix for a bub. There may be occassions when a CA only partly completes the HDC refusal task. In these cases the message will display a colon with nothing to its right hand side. Fix implemented to remove the colon if there is no reason provided for CA's HDC refusal.

* DCS-411: code refactor following PR comments

Co-authored-by: G <gurnank@outlook.com>