DCS-471 CAs HDC refusal is not recorded if reason not provided
If CA refuses HDC by selecting Yes in the page /hdc/finalChecks/refuse/{bookingId} but doesn't follow it up with a reason in page /hdc/finalChecks/refusal/{bookingId}, then when CA returns back to the CA tasklist, the previous Yes decision is deleted because the refusal has not been fully completed. The refusal can not be treated as complete without a reason being registered. 
The tasklist will not therefore display any refusal message and will have returned to the original state.