DCS-609 Adding fallback to load RO caseload (#1221)
Currently to build up the RO caselist, we look up offender data using a staff code

1.) We initially search the local staff table for a mapping from auth user to delius staff code
2.) If that fails, we make a request to delius to find the staff code for the currently logged in user.

After the probation area split on the weekend, new staff codes will be generated so our mappings in 1.) will
become out of date. We hoped to migrate these users to use delius credentials to access these services but
unfortunately have not had time to organise comms.

To get round this, this change will store the users currently mapped delius username in the local db.
It will then use this as a fallback when 1.) and 2.) fail to allow retrieval of an appropriate staff code
which can then be used to retrieve additional data.