DCS-631 removed inappropriate warning message (#1239)
The ticket was for the removal of the warning message “If any of the above details have been changed please make sure the prison case admin is notified to ensure a new licence is created if applicable”. 
The warning message was appearing in the screen seen by the RO when attempting to vary a licence’s conditions. Note, a vary can only be done once a prisoner has been released. The message was not appearing at any stage of licence creation prior to the prisoner being released. 
Once a prisoner has left prison, the only officials that deals with a prisoner are ROs. CAs and DMs are no longer relevant because they only deal with in-prison cases

The effect of the above is, the message shouldn’t be displayed anywhere, either before or after prison release. So I have deleted server/views/includes/roModificationWarning.pug (as it was the small file that defined the warning message) and deleted any references to it application-wide.

Co-authored-by: G <gurnank@outlook.com>